JAVA-2843: Move writeConcernErrorTransformer method
Move from WriteConcernHelper to CommandOperationHelper to aid in a
subsequent move of WriteConcernHelper to an internal package
